feat: update docs; restructure TS files
