+++ src/parsers/custom-parser-service.interface.ts
